NOISSUE: simplify graph output (remove console image gen)
