+++ Gopkg.toml
@@ -108,11 +108,3 @@ required = [
 [[constraint]]
   name = "github.com/cheggaaa/pb"
   version = "3.0.1"
-
-[[constraint]]
-  name = "github.com/campoy/tools"
-  version = "2.0.1"
-
-[[constraint]]
-  name = "github.com/wcharczuk/go-chart"
-  version = "2.0.1"
